requirements.txt
streamlit
groq
langchain
langchain-community
langchain-groq
langgraph
chromadb
pymupdf
pdfplumber
paddleocr
pillow
opencv-python
sentence-transformers
sqlalchemy
pandas
python-dotenv
bcrypt
pyjwt
fastapi
uvicorn
plotly
pydantic
python-multipart
tqdm
numpy
torch
torchvision
paddlepaddle
